Make pex_logging RFC-229 compliant.
