+++ .gitignore
@@ -19,22 +19,22 @@ version.py
 examples/debug
 examples/log
 examples/trace
-tests/testBlockTimingLog
-tests/testDebug
-tests/testDebug_results.txt
-tests/testDefLog
-tests/testFileDest
+tests/test_blockTimingLog
+tests/test_debug
+tests/test_debug_results.txt
+tests/test_defLog
+tests/test_fileDest
 tests/testFileDestination-out.txt
-tests/testLog
-tests/testLogDestination
-tests/testLogDestination_results.txt
-tests/testLogFormatter
-tests/testLogRecord
-tests/testLog_results.txt
-tests/testNoDebug
-tests/testNoDebug_results.txt
-tests/testNoTrace
-tests/testPropertyPrinter
-tests/testThresholdMemory
-tests/testTrace
-tests/timeSyscalls
+tests/test_log
+tests/test_logDestination
+tests/test_logDestination_results.txt
+tests/test_logFormatter
+tests/test_logRecord
+tests/test_log_results.txt
+tests/test_noDebug
+tests/test_noDebug_results.txt
+tests/test_noTrace
+tests/test_propertyPrinter
+tests/test_thresholdMemory
+tests/test_trace
+tests/test_timeSyscalls
