+++ tests/test_defLog.cc
