+++ tests/test_log.cc
