+++ tests/test_log.py
@@ -21,7 +21,7 @@
 #
 
 #
-# A wrapper test script around testLogDestination
+# A wrapper test script around testLog
 #
 # The wrapped test program writes output to the screen.  This script tests
 # the output by redirecting it to a file and comparing with canonical
@@ -41,10 +41,10 @@ import verifyOutput  # noqa E402 module level import not at top of file
 sys.path = origpath
 
 
-class TestLogDestination(lsst.utils.tests.TestCase):
+class TestLog(lsst.utils.tests.TestCase):
 
-    def testLogDestination(self):
-        self.assertLogs("testLogDestination", workdir=os.path.dirname(__file__))
+    def testLog(self):
+        self.assertLogs("test_log", workdir=os.path.dirname(__file__))
 
 
 class TestMemory(lsst.utils.tests.MemoryTestCase):
