+++ tests/test_logDestination.cc
