+++ tests/test_logFormatter.cc
