+++ tests/test_logRecord.cc
